# [Bug](function) make VcompoundPred optimization work well