--- a/be/src/vec/exprs/vcompound_pred.h
+++ b/be/src/vec/exprs/vcompound_pred.h
@@ -57,7 +57,10 @@ public:
     Status execute(VExprContext* context, vectorized::Block* block,
                    int* result_column_id) override {
         if (children().size() == 1 || !_all_child_is_compound_and_not_const() ||
-            _children[0]->is_nullable() != _children[1]->is_nullable()) {
+            _children[0]->is_nullable() || _children[1]->is_nullable()) {
+            // TODO:
+            // When the child is nullable, make the optimization also take effect, and the processing of this piece may be more complicated
+            // https://dev.mysql.com/doc/refman/8.0/en/logical-operators.html
             return VectorizedFnCall::execute(context, block, result_column_id);
         }
 
@@ -167,25 +170,12 @@ private:
 
     uint8* _get_raw_data(ColumnPtr column) const {
         if (column->is_nullable()) {
-            // true and null = null
-            // false and null = false
-            // true or null = true
-            // false or null = null
-            // For simplicity, we directly treat null as false, this logic only used for compound predicate.
-            size_t size = column->size();
-            uint8* null_map = _get_null_map(column);
-            uint8* data = assert_cast<ColumnUInt8*>(
-                                  assert_cast<ColumnNullable*>(column->assume_mutable().get())
-                                          ->get_nested_column_ptr()
-                                          .get())
-                                  ->get_data()
-                                  .data();
-
-            for (size_t i = 0; i < size; i++) {
-                data[i] &= !null_map[i];
-                null_map[i] = 0;
-            }
-            return data;
+            return assert_cast<ColumnUInt8*>(
+                           assert_cast<ColumnNullable*>(column->assume_mutable().get())
+                                   ->get_nested_column_ptr()
+                                   .get())
+                    ->get_data()
+                    .data();
         } else {
             return assert_cast<ColumnUInt8*>(column->assume_mutable().get())->get_data().data();
         }