@BiteTheDDDDt 2023-05-22
run buildall